backend/requirements.txt
fastapi==0.127.1
uvicorn[standard]
duckdb==1.4.3
pydantic==2.12.5
python-dotenv==1.2.1